RISCV/ComRV: Add handling for indirect overlay calls
When loading the address of a function pointer to an overlay call,
we instead load the address of the overlay group.

When making an indirect call from a non-overlay function, we now
emit some code to first check whether the function being called
is an overlay function before calling into the overlay system,
otherwise the function is just called directly.