ComRV: Disable X28-X31 for FastCC
This unconditionally disables X28-X31 in FastCC, since these might be
reserved. This is safe since FastCC is only used internally for static
functions whose addresses are never taken.
This patch will be replaced at some point with a better method of
recalculating the available registers for FastCC based on those which
have been explicitly reserved.