[RISCV] Add pre-emit pass to optimize multiple uses of uncompressible registers
In the case where multiple instructions use the value of an
uncompressible register, and if these instructions would otherwise be
compressed, it makes sense in terms of code size to add a copy to a
compressible register if there is one available, and replace the uses
with this compressible register instead such that code size is reduced
when these instructions are compressed.

This patch adds a pass which detects cases of this on the basic block
level, and performs the transformation. The logic necessary for
detecting the case where SW/LW instruction may be compressed is
implemented but logic for other instructions could easily be added in
the same manner.