[RISCV] Allow tail call optimisation to occur in more places.
Consider LLVM IR like this:

  define dso_local i32 @some_func() local_unnamed_addr #0 {
  entry:
    %call = tail call zeroext i8 @other_func() #2
    %conv = zext i8 %call to i32
    ret i32 %conv
  }

The function 'other_func' is returning an i8, while some_func is
returning i32.  In order to create the i32 return value from the i8
returned by some_func a zext instruction is added.

However, for RISC-V the ABI states that the i8 return value will be
either zero or sign extended to xlen, which in this case is i32.  In
effect the zext is free.

Indeed, we see this in the final code generation, the RISC-V backend
does not produce any assembler instructions for the zext instruction.

The problem though is that the presence if this zext in the IR
prevents the tail call optimisation from occurring.

This commit attempts to solve this problem and allow the tail call
optimisation to occur.

The core of this work is to implement TargetLowering::isExtFreeImpl
for the RISC-V target.  Calls to this function are then added in a
couple of places so that we can detect when the zext instruction is
not going to generate any assembler code, and thus we could perform a
tail-call.

A similar situation could occur but with sign extensions not zero
extensions, this should be covered too by this patch.